Added a select school field for create new user... can't get the destroy method to require the user be from the school he is trying to edit

+++ a_parametric_new_school/app/controllers/schools_controller.rb
@@ -56,10 +56,14 @@ class SchoolsController < ApplicationController
 
   def destroy
     school = School.find(params[:id])
-    user = User.find_by(email: params[:email])
-
-    if
-    school.destroy
+    if session[:user_id].school_id == user.school_id
+      school.destroy
+    else
+      flash[:alert] = "Could not be edited due to errors."
+      render locals: {
+        school: school
+      }
+    end
     redirect_to root_path
   end
 end
